+++ src/main.rs
@@ -1,96 +1,17 @@
+mod handlers;
+pub mod repositories;
+
 use axum::extract::Extension;
-use serde::{Deserialize, Serialize};
+use handlers::create_todo;
+use repositories::{TodoRepository, TodoRepositoryForMemory};
+use std::env;
 use std::net::SocketAddr;
-use std::sync::{Arc, RwLock};
-use std::{collections::HashMap, env};
-use tower::util::Optional;
+use std::sync::Arc;
 
 use axum::{
-    http::StatusCode,
-    response::IntoResponse,
     routing::{get, post},
-    Json, Router,
+    Router,
 };
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-enum RepositoryError {
-    #[error("Not Found, id is {0}")]
-    NotFound(i32),
-}
-
-pub trait TodoRepository: Clone + std::marker::Send + std::marker::Sync + 'static {
-    fn create(&self, payload: CreateTodo) -> Todo;
-    fn find(&self, id: i32) -> Option<Todo>;
-    fn all(&self) -> Vec<Todo>;
-    fn update(&self, id: i32, payload: UpdateTodo) -> anyhow::Result<Todo>;
-    fn delete(&self, id: i32) -> anyhow::Result<()>;
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
-pub struct Todo {
-    id: i32,
-    text: String,
-    completed: bool,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
-pub struct CreateTodo {
-    text: String,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
-pub struct UpdateTodo {
-    text: Option<String>,
-    completed: Option<bool>,
-}
-
-impl Todo {
-    pub fn new(id: i32, text: String) -> Self {
-        Self {
-            id,
-            text,
-            completed: false,
-        }
-    }
-}
-
-type TodoData = HashMap<i32, Todo>;
-
-#[derive(Clone, Debug)]
-pub struct TodoRepositoryForMemory {
-    store: Arc<RwLock<TodoData>>,
-}
-
-impl TodoRepositoryForMemory {
-    pub fn new() -> Self {
-        TodoRepositoryForMemory {
-            store: Arc::default(),
-        }
-    }
-}
-
-impl TodoRepository for TodoRepositoryForMemory {
-    fn create(&self, payload: CreateTodo) -> Todo {
-        todo!();
-    }
-
-    fn find(&self, id: i32) -> Option<Todo> {
-        todo!();
-    }
-
-    fn all(&self) -> Vec<Todo> {
-        todo!();
-    }
-
-    fn update(&self, id: i32, payload: UpdateTodo) -> anyhow::Result<Todo> {
-        todo!();
-    }
-
-    fn delete(&self, id: i32) -> anyhow::Result<()> {
-        todo!();
-    }
-}
 
 #[tokio::main]
 async fn main() {
@@ -121,22 +42,10 @@ async fn root() -> &'static str {
     "Hello, world!"
 }
 
-pub async fn create_todo<T: TodoRepository>(
-    Json(payload): Json<CreateTodo>,
-    Extension(repository): Extension<Arc<T>>,
-) -> impl IntoResponse {
-    let todo = repository.create(payload);
-
-    (StatusCode::CREATED, Json(todo))
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
-    use axum::{
-        body::Body,
-        http::{header, Method, Request},
-    };
+    use axum::{body::Body, http::Request};
     use tower::ServiceExt;
 
     #[tokio::test]
